+++ server/src/main/java/org/opensearch/action/search/SearchRequestStats.java
@@ -38,14 +38,12 @@ public final class SearchRequestStats extends SearchRequestOperationsListener {
 
     @Inject
     public SearchRequestStats(
-        ClusterService clusterService,
-        SearchRequestListenerManager searchRequestListenerManager
+        ClusterService clusterService
     ) {
         clusterService.getClusterSettings().addSettingsUpdateConsumer(SEARCH_REQUEST_STATS_ENABLED, this::setEnabled);
         for (SearchPhaseName searchPhaseName : SearchPhaseName.values()) {
             phaseStatsMap.put(searchPhaseName, new StatsHolder());
         }
-        this.searchRequestListenerManager = searchRequestListenerManager;
     }
 
     public long getPhaseCurrent(SearchPhaseName searchPhaseName) {